filtered to just saved alignments
